+++ lib/Model/Tag.php
@@ -16,7 +16,7 @@ use PhpBench\Storage\Exception\InvalidTagException;
 
 final class Tag
 {
-    public const REGEX_PATTERN = '\\w+';
+    public const REGEX_PATTERN = '[\\w\.]+';
 
     /**
      * @var string
@@ -25,9 +25,13 @@ final class Tag
 
     public function __construct(string $tag)
     {
+        // be restrictive with tag chars as:
+        //
+        // - `-` is reserved currently (e.g. my-tag-5 will show the 5th instance of my-tag)
+        // - we don't know how tags will be used in storage implementations
         if (!preg_match(sprintf('/^%s$/', self::REGEX_PATTERN), $tag)) {
             throw new InvalidTagException(sprintf(
-                'Tag must be non-empty string of alphanumeric characters and _, got "%s"',
+                'Tag must be non-empty string of alphanumeric characters, "." or "_". Got "%s"',
                 $tag
             ));
         }